Fix bug where a shortcut containing ctrl toggled the context menu
Pass for macOS Legacy 1.0 introduced the ability to open the context menu with
ctrl-click. Unfortunately, the code was only checking if ctrl was pressed, but
not if left-click was used at the same time. This resulted in the problem that
the context menu was opened instead of the popover if the shortcut contained
ctrl.